# social-flow-backend/openapi.yaml
openapi: 3.0.3
info:
  title: Social Flow Backend API
  description: |
    A comprehensive social media backend API combining YouTube and Twitter features
    with advanced AI/ML capabilities. This API supports video streaming, social posts,
    real-time interactions, content moderation, analytics, and machine learning features.
  version: 1.0.0
  contact:
    name: Social Flow Team
    email: support@socialflow.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8000/api/v1
    description: Development server
  - url: https://api.socialflow.com/api/v1
    description: Production server
security:
  - BearerAuth: []
  - ApiKeyAuth: []
paths:
...
  /videos/upload:
    post:
      tags:
        - videos
      summary: Upload a video
      description: Upload a new video file for processing and streaming
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
                  description: Video file to upload
                title:
                  type: string
                  description: Video title
                description:
                  type: string
                  description: Video description
                tags:
                  type: string
                  description: Comma-separated tags
              required:
                - file
                - title
      responses:
        '200':
          description: Video uploaded successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  video_id:
                    type: string
                    format: uuid
                  status:
                    type: string
                    enum: [uploaded, processing, processed, failed]
                  message:
                    type: string
        '400':
          description: Bad request - invalid file or data
        '401':
          description: Unauthorized
  /videos/{video_id}:
    get:
      tags:
        - videos
      summary: Get video details
      description: Retrieve video information and metadata
      parameters:
        - name: video_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Video details retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Video'
        '404':
          description: Video not found
  /videos/{video_id}/stream:
    get:
      tags:
        - videos
      summary: Get video stream URL
      description: Get streaming URL for video playback
      parameters:
        - name: video_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: quality
          in: query
          schema:
            type: string
            enum: [auto, 240p, 360p, 480p, 720p, 1080p]
            default: auto
      responses:
        '200':
          description: Streaming URL retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  video_id:
                    type: string
                    format: uuid
                  streaming_url:
                    type: string
                    format: uri
                  quality:
                    type: string
                  duration:
                    type: number
                  resolution:
                    type: string
        '404':
          description: Video not found
  /videos/{video_id}/like:
    post:
      tags:
        - videos
      summary: Like a video
      description: Like a video
      security:
        - BearerAuth: []
      parameters:
        - name: video_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Video liked successfully
        '401':
          description: Unauthorized
        '404':
          description: Video not found
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
  schemas:
...
    Video:
      type: object
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        description:
          type: string
        tags:
          type: string
        filename:
          type: string
        file_size:
          type: integer
        duration:
          type: number
        resolution:
          type: string
        bitrate:
          type: integer
        codec:
          type: string
        thumbnail_url:
          type: string
          format: uri
        hls_url:
          type: string
          format: uri
        dash_url:
          type: string
          format: uri
        streaming_url:
          type: string
          format: uri
        status:
          type: string
          enum: [uploading, processing, processed, failed, deleted]
        visibility:
          type: string
          enum: [public, unlisted, private]
        is_approved:
          type: boolean
        is_flagged:
          type: boolean
        is_rejected:
          type: boolean
        views_count:
          type: integer
        likes_count:
          type: integer
        dislikes_count:
          type: integer
        comments_count:
          type: integer
        shares_count:
          type: integer
        total_watch_time:
          type: number
        average_watch_time:
          type: number
        retention_rate:
          type: number
        is_monetized:
          type: boolean
        ad_revenue:
          type: number
        engagement_rate:
          type: number
        like_ratio:
          type: number
        owner_id:
          type: string
          format: uuid
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...